move staticData to store
