+++ components/core/Autocomplete/Autocomplete.tsx
@@ -11,10 +11,7 @@ import {
   StyledListboxItem,
 } from './styles/autocomplete';
 
-import {
-  PlaceholderText,
-  DEFAULT_TEXT_LOADING,
-} from '../../../helpers/staticData';
+import { PlaceholderText, DEFAULT_TEXT_LOADING } from '../../../store/statics';
 import { Restaurant, CardType } from '../../../helpers/types';
 
 export type AutocompleteProps = {
