+++ components/layouts/Layout/Layout.tsx
@@ -11,7 +11,7 @@ import { StyledLayout, StyledMain, StyledScrollUpButton } from './styles';
 import {
   SCROLL_DELAY,
   SHOWING_SCROLLUP_BUTTON_HEIGHT,
-} from '../../../helpers/staticData';
+} from '../../../store/statics';
 
 type LayoutProps = {
   children: ReactNode;
