+++ components/ui/ReviewCard/ReviewCard.tsx
@@ -12,7 +12,7 @@ import {
   StyledReviewText,
 } from './styles';
 
-import { THUMB_GENERIC_SRC } from '../../../helpers/staticData';
+import { THUMB_GENERIC_SRC } from '../../../store/statics';
 
 type ReviewCardProps = {
   userImgSrc: string;
