+++ pages/_app.tsx
@@ -9,7 +9,7 @@ import { throttle } from 'lodash';
 import appReducer from '@store/reducer';
 
 import { InitialAppState } from '@helpers/types';
-import { DEFAULT_BREADCRUMB } from '@helpers/staticData';
+import { DEFAULT_BREADCRUMB } from '@store/statics';
 import '@helpers/Grid/Grid.scss';
 import { GlobalStyles } from '@helpers/GlobalStylesHelper';
 import { theme } from '@helpers/Theme';
